add MySQL compatibility
